# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Romanian Festival Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /dashboard/summary:
    get:
      operationId: getDashboardSummary
      tags: [dashboard]
      summary: Get current year dashboard summary
      security:
        - clerkAuth: []
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"
  /dashboard/financials:
    get:
      operationId: getDashboardFinancials
      tags: [dashboard]
      summary: Get financial breakdown for current year
      security:
        - clerkAuth: []
      parameters:
        - name: yearId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Financials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FinancialSummary"
  /dashboard/activity:
    get:
      operationId: getRecentActivity
      tags: [dashboard]
      summary: Get recent activity feed
      security:
        - clerkAuth: []
      responses:
        "200":
          description: Activity feed
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityItem"

  # ─── FESTIVAL YEARS ──────────────────────────────────────
  /festival-years:
    get:
      operationId: listFestivalYears
      tags: [festival-years]
      summary: List all festival years
      security:
        - clerkAuth: []
      responses:
        "200":
          description: Festival years
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FestivalYear"

    post:
      operationId: createFestivalYear
      tags: [festival-years]
      summary: Create a new festival year
      security:
        - clerkAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FestivalYearInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FestivalYear"
  /festival-years/{id}:
    patch:
      operationId: updateFestivalYear
      tags: [festival-years]
      summary: Update a festival year
      security:
        - clerkAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FestivalYearUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FestivalYear"

  # ─── SETTINGS ────────────────────────────────────────────
...
  /export/vendors:
    get:
      operationId: exportVendors
      tags: [export]
      summary: Export vendors as Excel file
      security:
        - clerkAuth: []
      parameters:
        - name: yearId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Excel file
          content:
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
              schema:
                type: string
                format: binary
  /export/sponsors:
    get:
      operationId: exportSponsors
      tags: [export]
      summary: Export sponsors as Excel file
      security:
        - clerkAuth: []
      parameters:
        - name: yearId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Excel file
          content:
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
              schema:
                type: string
                format: binary
  /export/volunteers:
    get:
      operationId: exportVolunteers
      tags: [export]
      summary: Export volunteers as Excel file
      security:
        - clerkAuth: []
      parameters:
        - name: yearId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Excel file
          content:
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
              schema:
                type: string
                format: binary
components:
  securitySchemes:
    clerkAuth:
      type: http
      scheme: bearer
  schemas:
...
    FestivalYear:
      type: object
      required: [id, year, eventName, eventDate, isActive, createdAt]
      properties:
        id:
          type: integer
        year:
          type: integer
        eventName:
          type: string
        eventDate:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
    FestivalYearInput:
      type: object
      required: [year, eventName, eventDate]
      properties:
        year:
          type: integer
        eventName:
          type: string
        eventDate:
          type: string
        isActive:
          type: boolean
    FestivalYearUpdate:
      type: object
      properties:
        eventName:
          type: string
        eventDate:
          type: string
        isActive:
          type: boolean
...
    DashboardSummary:
      type: object
      required: [festivalYear, countdown, vendorStats, sponsorStats, volunteerStats, totalRevenue, pendingActions]
      properties:
        festivalYear:
          $ref: "#/components/schemas/FestivalYear"
        countdown:
          type: integer
          description: Days until the festival
        vendorStats:
          $ref: "#/components/schemas/ApplicationStats"
        sponsorStats:
          $ref: "#/components/schemas/ApplicationStats"
        volunteerStats:
          $ref: "#/components/schemas/ApplicationStats"
        totalRevenue:
          type: number
        pendingActions:
          type: integer
          description: Total applications needing attention
    ApplicationStats:
      type: object
      required: [total, pending, approved, rejected, paid, finalApproved]
      properties:
        total:
          type: integer
        pending:
          type: integer
        approved:
          type: integer
        rejected:
          type: integer
        paid:
          type: integer
        finalApproved:
          type: integer
    FinancialSummary:
      type: object
      required: [vendorRevenue, sponsorRevenue, totalRevenue, vendorCount, sponsorCount]
      properties:
        vendorRevenue:
          type: number
        sponsorRevenue:
          type: number
        totalRevenue:
          type: number
        vendorCount:
          type: integer
        sponsorCount:
          type: integer
        recentPayments:
          type: array
          items:
            $ref: "#/components/schemas/PaymentRecord"
    PaymentRecord:
      type: object
      required: [type, name, amount, paidAt]
      properties:
        type:
          type: string
          enum: [vendor, sponsor]
        name:
          type: string
        amount:
          type: number
        paidAt:
          type: string
    ActivityItem:
      type: object
      required: [id, type, message, entityType, createdAt]
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [new_application, approved, rejected, paid, final_approved, assigned]
        message:
          type: string
        entityType:
          type: string
          enum: [vendor, sponsor, volunteer]
        entityId:
          type: integer
        createdAt:
          type: string
...