# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Romanian Festival Management API
 * OpenAPI spec version: 0.1.0
 */
import type { ApplicationStats } from './applicationStats';
import type { FestivalYear } from './festivalYear';

export interface DashboardSummary {
  festivalYear: FestivalYear;
  /** Days until the festival */
  countdown: number;
  vendorStats: ApplicationStats;
  sponsorStats: ApplicationStats;
  volunteerStats: ApplicationStats;
  totalRevenue: number;
  /** Total applications needing attention */
  pendingActions: number;
}